+++ app/Controllers/CommunityController.php
@@ -2,8 +2,6 @@
 
 namespace App\Controllers;
 
-use App\Models\CommunityModel;
-use App\Models\OrganizationModel;
 use CodeIgniter\HTTP\Response;
 use CodeIgniter\RESTful\ResourceController;
 use Codewrite\CoopAuth\ApiResponse;
@@ -12,22 +10,18 @@ class CommunityController extends ResourceController
 {
     protected $modelName = 'App\Models\CommunityModel';
     protected $format    = 'json';
+    protected $allowedColumns = [];
 
     public function index()
     {
-
-        $params = $this->request->getVar(['columns', 'sort', 'page', 'pageSize']);
-        $allowedColumns = [];
-
-        $response = new ApiResponse($this->model, $params, $allowedColumns);
+        $params = $this->request->getVar(['columns', 'filters', 'sort', 'page', 'pageSize']);
+        $response = new ApiResponse($this->model, $params, $this->allowedColumns);
 
         return $response->getCollectionResponse();
     }
 
     public function create()
     {
-        $data = (array)$this->request->getVar();
-
         $rules = config('Validation')->create['communities'];
         // Validate input
         if (!$this->validate($rules)) {
@@ -37,15 +31,13 @@ class CommunityController extends ResourceController
                 'error'   => $this->validator->getErrors()
             ], Response::HTTP_BAD_REQUEST);
         }
-        $orgModel = new OrganizationModel();
-
-        if (!$orgModel->find($data['orgid'])) {
-            return $this->respond([
-                'status'  => false,
-                'message' => 'Failed validating data',
-                'error'   =>    ["orgid" => "The Organization doesn't exist."]
-            ], Response::HTTP_BAD_REQUEST);
-        }
+        $data = $this->validator->getValidated();
+        $data['com_code'] = $data['com_code'] ?? $this->model->generateCode($data['orgid']);
+        $data['creator'] = auth()->user_id();
+        
+        $response = auth()->can('create', 'communities', ['owner', 'orgid'], [$data]);
+        if ($response->denied())
+            return $response->responsed();
 
         if ($this->model->save($data)) {
             return $this->respondCreated([
@@ -64,7 +56,16 @@ class CommunityController extends ResourceController
 
     public function update($id = null)
     {
-        $data = $this->request->getRawInput();
+        $community = $this->model->find($id);
+        if (!$community) {
+            return $this->respond([
+                'status' => false,
+                'message' => 'Community not found'
+            ], Response::HTTP_NOT_FOUND);
+        }
+        $response = auth()->can('update', 'communities', ['owner', 'orgid'], [$community]);
+        if ($response->denied())
+            return $response->responsed();
 
         $rules = config('Validation')->update['communities'];
         // Validate input
@@ -75,13 +76,7 @@ class CommunityController extends ResourceController
                 'error'   => $this->validator->getErrors()
             ], Response::HTTP_BAD_REQUEST);
         }
-
-        if (!$this->model->find($id)) {
-            return $this->respond([
-                'status' => false,
-                'message' => 'Community not found'
-            ], Response::HTTP_NOT_FOUND);
-        }
+        $data = $this->validator->getValidated();
 
         $this->model->update($id, $data);
         return $this->respond([
@@ -93,12 +88,10 @@ class CommunityController extends ResourceController
 
     public function show($id = null)
     {
-        $params = $this->request->getVar(['columns', 'sort', 'page', 'pageSize']);
-        $allowedColumns = [];
-        $this->model->find($id);
-        $response = new ApiResponse($this->model, $params, $allowedColumns);
+        $params = $this->request->getVar(['columns']);
+        $response = new ApiResponse($this->model, $params, $this->allowedColumns);
 
-        return $response->getSingleResponse();
+        return $response->getSingleResponse(true, ['owner', 'orgid']);
     }
 
     public function delete($id = null)
@@ -110,6 +103,10 @@ class CommunityController extends ResourceController
                 'message' => 'Community not found'
             ], Response::HTTP_NOT_FOUND);
         }
+
+        $response = auth()->can('delete', 'communities', ['owner', 'orgid'], [$community]);
+        if ($response->denied())
+            return $response->responsed();
 
         if ($this->model->delete($id)) {
             return $this->respondDeleted([
@@ -125,52 +122,52 @@ class CommunityController extends ResourceController
         }
     }
 
-     // Pull changes from the server
-     public function pull()
-     {
-         $lastSyncTime = $this->request->getGet('lastSyncTime');
- 
-         // Fetch updated after the last pulled timestamp
-         $records = $this->model
-             ->where('updated_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
-             ->findAll();
-         $deletedRecords = $this->model->select(['id', 'deleted_at'])
-             ->where('deleted_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
-             ->onlyDeleted()->findAll();
- 
-         return $this->respond([
-             'updated' => $records,
-             'deleted' => $deletedRecords,
-             'timestamp' =>  date('Y-m-d H:i:s', strtotime('now')) // Current server time for synchronization
-         ]);
-     }
- 
-     // Push changes to the server
-     public function push()
-     {
-         $rules = config('Validation')->sync;
-         // Validate input
-         if (!$this->validate($rules)) {
-             return $this->respond([
-                 'status'  => false,
-                 'message' => 'Failed validating data',
-                 'error'   => $this->validator->getErrors()
-             ], Response::HTTP_BAD_REQUEST);
-         }
- 
-         $updates = $this->request->getVar('updated');
-         $nrowsUpdated = sizeof($updates);
-         $deleted = $this->request->getVar('deleted');
-         $nrowsDeleted = sizeof($updates);
- 
-         if ($nrowsUpdated > 0)
-             $this->model->builder()->updateBatch($updates, ['id'], sizeof($updates));
-         if ($nrowsDeleted > 0)
-             $this->model->builder()->updateBatch($deleted, ['id'], sizeof($deleted));
- 
-         return $this->respond([
-             'status' => true,
-             'message' => 'Sync completed successfully'
-         ], Response::HTTP_OK);
-     }
+    // Pull changes from the server
+    public function pull()
+    {
+        $lastSyncTime = $this->request->getGet('lastSyncTime');
+
+        // Fetch updated after the last pulled timestamp
+        $records = $this->model
+            ->where('updated_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
+            ->findAll();
+        $deletedRecords = $this->model->select(['id', 'deleted_at'])
+            ->where('deleted_at >', date('Y-m-d H:i:s', strtotime($lastSyncTime)))
+            ->onlyDeleted()->findAll();
+
+        return $this->respond([
+            'updated' => $records,
+            'deleted' => $deletedRecords,
+            'timestamp' =>  date('Y-m-d H:i:s', strtotime('now')) // Current server time for synchronization
+        ]);
+    }
+
+    // Push changes to the server
+    public function push()
+    {
+        $rules = config('Validation')->sync;
+        // Validate input
+        if (!$this->validate($rules)) {
+            return $this->respond([
+                'status'  => false,
+                'message' => 'Failed validating data',
+                'error'   => $this->validator->getErrors()
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $updates = $this->request->getVar('updated');
+        $nrowsUpdated = sizeof($updates);
+        $deleted = $this->request->getVar('deleted');
+        $nrowsDeleted = sizeof($updates);
+
+        if ($nrowsUpdated > 0)
+            $this->model->builder()->updateBatch($updates, ['id'], sizeof($updates));
+        if ($nrowsDeleted > 0)
+            $this->model->builder()->updateBatch($deleted, ['id'], sizeof($deleted));
+
+        return $this->respond([
+            'status' => true,
+            'message' => 'Sync completed successfully'
+        ], Response::HTTP_OK);
+    }
 }
